outsourced header background to separate component
